Fixed missing file bug

+++ setup.py
@@ -8,12 +8,16 @@ coc_settings_dir = Path("~/.config/nvim").expanduser()
 vimrc = vimrc_dir / ".vimrc"
 coc_settings = coc_settings_dir / "coc-settings.json"
 
+vimrc_exists = False
+coc_settings_exists = False
+
 if not os.path.exists(vimrc):
     print("No already existing vimrc found")
     with open(vimrc, "w"):
         pass
 else:
     print("Already existing vimrc found")
+    vimrc_exists = True
     shutil.copyfile(vimrc, vimrc_dir / ".vimrc.orig")
     print("Original vimrc --> .vimrc.orig")
 
@@ -24,6 +28,7 @@ if not os.path.exists(coc_settings):
         pass
 else:
     print("Already existing coc-settings.json found")
+    coc_settings_exists = True
     shutil.copyfile(coc_settings, coc_settings_dir / "coc-settings.json.orig")
     print("Original coc-settings.json --> coc-settings.json.orig")
 
@@ -33,7 +38,9 @@ shutil.copyfile("coc-settings.json", coc_settings)
 to_delete_orig = input("Do you want to delete original files? y/[n]: ")
 
 if to_delete_orig == "y":
-    os.remove(vimrc_dir / ".vimrc.orig")
-    os.remove(coc_settings_dir / "coc-settings.json.orig")
+    if vimrc_exists:
+        os.remove(vimrc_dir / ".vimrc.orig")
+    if coc_settings_exists:
+        os.remove(coc_settings_dir / "coc-settings.json.orig")
 
 print("Vim setup done!")
